improving visual difference between melee and explosive roomba
I lowered the sensor piece of the melee roomba. I noticed in gameplay that they're too difficult to differentiate because of the speed they move and how low-profile they are. Now you should have a clearer visual that one with a tall sensor is an exploding roomba and one with a low one is a melee roomba.